Rename target from rviz2 to schedule_visualizer. Update visualizer launch file
